+++ packages/api/src/modules/finance/types.ts
@@ -1,6 +1,6 @@
-import type { OnePaymentUseCase } from '@/modules/finance/application/use-cases/payment/one-payment.use-case'
-import type { PaymentHookUseCase } from '@/modules/finance/application/use-cases/payment/payment-hook.use-case'
-import type { AllTransactionsByUserUseCase } from '@/modules/finance/application/use-cases/transaction/all-by-user.use-case'
+import type { AllTransactionsByUserUseCase } from '@/modules/finance/application/use-cases/all-transactions-by-user.use-case'
+import type { OnePaymentUseCase } from '@/modules/finance/application/use-cases/one-payment.use-case'
+import type { PaymentHookUseCase } from '@/modules/finance/application/use-cases/payment-hook.use-case'
 
 export interface UseCases {
   payment: {
@@ -12,12 +12,10 @@ export interface UseCases {
   }
 }
 
-// Payment DTOs
-export { OnePaymentDto } from '@/modules/finance/application/dtos/payment/one-payment.dto'
-export { PaymentHookDto } from '@/modules/finance/application/dtos/payment/payment-hook.dto'
-
-// Transaction DTOs
-export { AllTransactionsByUserDto } from '@/modules/finance/application/dtos/transaction/all-by-user.dto'
+// DTOs
+export { OnePaymentDto } from '@/modules/finance/application/dtos/one-payment.dto'
+export { PaymentHookDto } from '@/modules/finance/application/dtos/payment-hook.dto'
+export { AllTransactionsByUserDto } from '@/modules/finance/application/dtos/all-transactions-by-user.dto'
 
 // Entities
 export { PaymentEntity } from '@/modules/finance/domain/entities/payment.entity'