+++ packages/api/src/modules/merchant/domain/entities/vendor-balance.entity.ts
@@ -1,7 +1,7 @@
 import { AbstractEntity } from '@/shared/abstracts/abstract.entity'
 
 export class VendorBalanceEntity extends AbstractEntity<VendorBalanceEntity> {
-  declare public balance: number
+  declare public balance: string
 
   declare public vendorId: string
 }